Suppressed Xcode compiler warnings.
Added initialization of some variables. This is not a real uninitialized
variable problem - when reading/writing valid charstrings, these
variables would all have been initialized before being used, but the
compiler can't tell that.

+++ libpsautohint/src/read.c
@@ -680,7 +680,7 @@ ParseString(const ACFontInfo* fontinfo, const char* s)
                         if (c == 0)
                             break;
                     }
-                    DoName(fontinfo, s0, s, s - s0 - 1);
+                    DoName(fontinfo, s0, s, (int)(s - s0 - 1));
                     if (c == '\0')
                         s--;
                     continue;
